rephrased documentation and split up tests for web and mobile platform.
